Refactor pick and place real robot code so that we can use it in sim (#168)
* improve pick tests and run experiments; add docs and flags

* code refactoring - moving things out of the grasp planner to make sure
we can test more stuff in simulation

* refactor grasping code

* finishing a refactor for the grasping code - pull motion plan generation
out of ROS code

* creating gras[ amd vpxe; gemeratpr' refacting code so it isnt in hw any
more

* cleaning up and addressing comments from the people

* add docstrings

* remove dead code
